requirements.dev.txt
coverage==3.7.1
django-nose==1.2
flake8==2.1.0
ipdb==0.7
ipython==0.13
nose==1.3.0
